Added audio code

+++ patent-service/src/service/summarization/patent_summarizer.py
@@ -46,30 +46,33 @@ class Pdf_Patent_Summarizer:
         """
         Summarizes the input text using BigBird-Pegasus.
         """
-        summary = ""
-        try:
-            logger.info(f"Generating summary of {pdf_file}")
-            inputs = self.preprocess_text(text)
-            outputs = self.model.generate(
-                inputs.input_ids,
-                max_length=200,  # Adjust based on desired summary length
-                min_length=50,
-                num_beams=4,
-                length_penalty=2.0,
-                early_stopping=True,
-                #callbacks = self.callbacks
-            )
-            summary = self.tokenizer.decode(outputs[0], skip_special_tokens=True)
-        except Exception as e:
-            logger.error(f"Error while generating summary of the PDF file ({pdf_file}): {str(e)}")
-        if summary:
-            with open(output_file, "w") as f:
-                f.write(summary)
-            logger.info(f"Generated summary of {pdf_file} at location {output_file}")  
-        #self.clearml_callback.flush_tracker(langchain_asset=self.model, name="sumarizer")      
+        if not utils.is_output_file_exists(output_file):
+            summary = ""
+            try:
+                logger.info(f"Generating summary of {pdf_file}")
+                inputs = self.preprocess_text(text)
+                outputs = self.model.generate(
+                    inputs.input_ids,
+                    max_length=200,  # Adjust based on desired summary length
+                    min_length=50,
+                    num_beams=4,
+                    length_penalty=2.0,
+                    early_stopping=True,
+                    #callbacks = self.callbacks
+                )
+                summary = self.tokenizer.decode(outputs[0], skip_special_tokens=True)
+            except Exception as e:
+                logger.error(f"Error while generating summary of the PDF file ({pdf_file}): {str(e)}")
+            if summary:
+                with open(output_file, "w") as f:
+                    f.write(summary)
+                logger.info(f"Generated summary of {pdf_file} at location {output_file}")  
+            #self.clearml_callback.flush_tracker(langchain_asset=self.model, name="sumarizer")
+        else:
+            logger.info(f"Saving processing!! Summary of {pdf_file} already exists at location {output_file}...")           
 
 patent_summarizer = Pdf_Patent_Summarizer()
-pdf_file = const.INPUT_PATENT_DIR_PATH + "/US10612929.pdf"
+pdf_file = const.INPUT_PATENT_DIR_PATH + "/US20190213407A1.pdf"
 _, file_name_without_ext=utils.get_file_name_and_without_extension(pdf_file)
 text = utils.get_pdf_text(pdf_file)
 output_file = const.OUTPUT_DIR + '/summary/'+file_name_without_ext
